+++ src/components/toast/_store/ToastProvider.tsx
@@ -1,0 +1,58 @@
+import React, { PropsWithChildren, useCallback, useMemo, useState } from 'react'
+import { Toast } from '../type'
+import { getRandomNumber } from '@/shared/utils/number'
+import { TOAST_TIMEOUT } from '../constant'
+import ToastPortal from '../ToastPortal'
+import useSafetyContext, { createSateyContext } from '@/hooks/useSafetyContext'
+
+type Value = {
+  showToast: ({ variety, desc }: Pick<Toast, 'desc' | 'variety'>) => void
+}
+const ToastContext = createSateyContext<Value>('ToastContext')
+
+function ToastProvider({ children }: PropsWithChildren) {
+  const [toasts, setToasts] = useState<Array<Toast>>([])
+
+  const clear = (id: number) => {
+    const timeout = setTimeout(() => {
+      setToasts((prev) => prev.filter((toast) => toast.id !== id))
+      clearTimeout(timeout)
+    }, TOAST_TIMEOUT)
+  }
+
+  const addToast = useCallback(
+    ({ variety, desc }: Pick<Toast, 'desc' | 'variety'>) => {
+      const id = getRandomNumber()
+
+      setToasts((prev) => [
+        ...prev,
+        {
+          id,
+          variety,
+          desc,
+        },
+      ])
+
+      clear(id)
+    },
+    []
+  )
+
+  const values = useMemo(
+    () => ({
+      showToast: addToast,
+    }),
+    [addToast]
+  )
+
+  return (
+    <ToastContext.Provider value={values}>
+      {children}
+      <ToastPortal toasts={toasts} />
+    </ToastContext.Provider>
+  )
+}
+
+export default ToastProvider
+
+export const useToastContext = () => useSafetyContext({ context: ToastContext })
