--- src/components/toast/useToast.tsx
@@ -1,38 +1,0 @@
-import { useCallback, useState } from 'react'
-
-import { getRandomNumber } from '@/shared/utils/number'
-import { Toast } from '@/components/toast/type'
-import { TOAST_TIMEOUT } from '@/components/toast/constant'
-
-const useToast = () => {
-  const [toasts, setToasts] = useState<Array<Toast>>([])
-
-  const clear = (id: number) => {
-    const timeout = setTimeout(() => {
-      setToasts((prev) => prev.filter((toast) => toast.id !== id))
-      clearTimeout(timeout)
-    }, TOAST_TIMEOUT)
-  }
-
-  const addToast = useCallback(
-    ({ variety, desc }: Pick<Toast, 'desc' | 'variety'>) => {
-      const id = getRandomNumber()
-
-      setToasts((prev) => [
-        ...prev,
-        {
-          id,
-          variety,
-          desc,
-        },
-      ])
-
-      clear(id)
-    },
-    []
-  )
-
-  return { addToast, toasts }
-}
-
-export default useToast
